finished detail screen

+++ app/src/main/java/com/arctouch/codechallenge/api/TmdbApi.kt
@@ -23,6 +23,9 @@ interface TmdbApi {
         @Query("language") language: String
     ): Observable<GenreResponse>
 
+    // REMOVING DEFAULT REGION FROM REQUEST BECAUSE WAS RETURNING ONLY 12 RESULTS
+    // MAKING IMPOSSIBLE TO TET INIFITY SCROLL
+
     @GET("movie/upcoming")
     fun upcomingMovies(
         @Query("api_key") apiKey: String,
@@ -32,7 +35,7 @@ interface TmdbApi {
 
     @GET("movie/{id}")
     fun movie(
-        @Path("id") id: Long,
+        @Path("id") id: Int,
         @Query("api_key") apiKey: String,
         @Query("language") language: String
     ): Observable<Movie>
